+++ pkg/controllers/clusterresourceplacement/controller.go
@@ -209,7 +209,7 @@ func (r *Reconciler) handleUpdate(ctx context.Context, crp *fleetv1beta1.Cluster
 
 	if r.UseNewConditions {
 		// We skip checking the last resource condition (available) because it will be covered by checking isRolloutCompleted func.
-		for i := RolloutStartedCondition; i < TotalCondition-1; i++ {
+		for i := condition.RolloutStartedCondition; i < condition.TotalCondition-1; i++ {
 			oldCond := oldCRP.GetCondition(string(i.ClusterResourcePlacementConditionType()))
 			newCond := crp.GetCondition(string(i.ClusterResourcePlacementConditionType()))
 			if !condition.IsConditionStatusTrue(oldCond, oldCRP.Generation) &&
@@ -1159,7 +1159,7 @@ func isRolloutCompleted(useNewConditions bool, crp *fleetv1beta1.ClusterResource
 	}
 
 	if useNewConditions {
-		for i := RolloutStartedCondition; i < TotalCondition; i++ {
+		for i := condition.RolloutStartedCondition; i < condition.TotalCondition; i++ {
 			if !condition.IsConditionStatusTrue(crp.GetCondition(string(i.ClusterResourcePlacementConditionType())), crp.Generation) {
 				return false
 			}
